Reworked the PALM analysis controller to handle errors caused in the worker threads.
Boost::thread does not provide a mechanism to allow exceptions thrown in other threads to propagate to the main thread, instead std::terminate() is called. This meant that, until now, whenever an exception was encountered in the PALM analysis, the program was terminated.

Now exceptions are handled with a kludgy workaround: PALMAnalysisController now has a new member, errorMessage, that is initialized to an empty string. When a worker thread catches an error it puts an error message in this member string. The main DoPALMAnalysis() periodically checks for this, and if found shuts down the worker threads an throws an ERROR_RUNNING_THREADED_ANALYSIS exception. This is caught in the main thread and should allow graceful exiting from the analysis.