Add a stringstream, analysisOptionsStream, that builds up a string with the relevant settings while parsing the options passed to ExecuteAnalyzePALMImages
The options used to analyze a given CCD file are very relevant to later analysis of the results. Previously the Igor procedure added the relevant information to the wave not as a key:value list. If outputting positions to a file is to be supported in the future then the XOP part should take over that functionality. For now this is done by incrementally adding to a stringstream while parsing the options.