Make FitPositions classes delete proposed particles if they do not yield a satisfactory localization.
The fitting of an emitter spot provides information on the shape of the spot,
which can help distinguish between a real emitter PSF and an artefact or two
overlapping emitters. By allowing localization algorithms to delete invalid spots
from the particle list these can be used not just for localization but also for
testing the vailidity of an emission spot.